Restructure the project to add 1.19-1.21 support as one jar.
