+++ .gitignore
@@ -1,4 +1,5 @@
 # Project exclude paths
 /.gradle/
-/build/
-/build/classes/java/main/
+/build
+/Core/build
+/v1_20_R4/build
