+++ Core/build.gradle
@@ -1,0 +1,29 @@
+plugins {
+    id("io.github.goooler.shadow") version "8.1.7"
+}
+
+dependencies {
+    compileOnly "io.papermc.paper:paper-api:1.19-R0.1-SNAPSHOT"
+    compileOnly "com.github.MilkBowl:VaultAPI:1.7"
+    compileOnly 'me.clip:placeholderapi:2.11.6'
+    compileOnly "net.kyori:adventure-api:4.17.0"
+    compileOnly "net.kyori:adventure-text-minimessage:4.17.0"
+    implementation "org.spongepowered:configurate-yaml:4.1.2"
+    implementation "org.bstats:bstats-bukkit:3.0.2"
+}
+
+java.toolchain {
+    JavaLanguageVersion.of(21)
+}
+
+shadowJar {
+    archiveClassifier.set(null)
+    relocate('org.spongepowered.configurate', 'com.github.lukesky19.skyshop.libs.configurate')
+    relocate('org.bstats', 'com.github.lukesky19.skyshop.libs.bstats')
+    minimize()
+    manifest {
+        attributes("paperweight-mappings-namespace": "mojang")
+    }
+}
+
+build.dependsOn shadowJar
