+++ Core/src/main/java/com/github/lukesky19/skyshop/configuration/settings/SettingsConfiguration.java
