+++ Core/src/main/java/com/github/lukesky19/skyshop/gui/MenuGUI.java
@@ -19,8 +19,8 @@ package com.github.lukesky19.skyshop.gui;
 
 import com.github.lukesky19.skyshop.SkyShop;
 import com.github.lukesky19.skyshop.configuration.locale.LocaleManager;
+import com.github.lukesky19.skyshop.configuration.menu.MenuConfiguration;
 import com.github.lukesky19.skyshop.configuration.menu.MenuManager;
-import com.github.lukesky19.skyshop.configuration.shop.ShopConfiguration;
 import com.github.lukesky19.skyshop.configuration.shop.ShopManager;
 import com.github.lukesky19.skyshop.util.enums.TransactionType;
 import com.github.lukesky19.skyshop.util.gui.InventoryButton;
@@ -29,8 +29,6 @@ import com.github.lukesky19.skyshop.util.gui.InventoryManager;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.format.TextDecoration;
 import net.kyori.adventure.text.minimessage.MiniMessage;
-import net.kyori.adventure.text.minimessage.tag.resolver.Placeholder;
-import net.kyori.adventure.text.minimessage.tag.resolver.TagResolver;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
@@ -42,17 +40,16 @@ import java.util.List;
 import java.util.Map;
 
 /**
- * This class is called to create a shop inventory for a player to access an individual shop category.
+ * This class is called to create a menu inventory for a player to access the shop.
 */
-public class ShopGUI extends InventoryGUI {
+public class MenuGUI extends InventoryGUI {
     final SkyShop skyShop;
     final MenuManager menuManager;
     final InventoryManager inventoryManager;
     final ShopManager shopManager;
     final LocaleManager localeManager;
-    final Map.Entry<String, ShopConfiguration.ShopPage> pageEntry;
-    final String shopId;
-    final int pageNum;
+    final Map.Entry<String, MenuConfiguration.MenuPage> pageEntry;
+    final Integer pageNum;
 
     /**
      * Constructor
@@ -63,23 +60,14 @@ public class ShopGUI extends InventoryGUI {
      * @param localeManager A LocaleManager instance.
      * @param pageEntry The page configuration to create the Inventory based on.
      * @param pageNum The page number associated with the Inventory being created.
-    */
-    public ShopGUI(
-            SkyShop skyShop,
-            MenuManager menuManager,
-            ShopManager shopManager,
-            InventoryManager inventoryManager,
-            LocaleManager localeManager,
-            Map.Entry<String, ShopConfiguration.ShopPage> pageEntry,
-            String shopId,
-            int pageNum) {
+     */
+    public MenuGUI(SkyShop skyShop, MenuManager menuManager, ShopManager shopManager, InventoryManager inventoryManager, LocaleManager localeManager, Map.Entry<String, MenuConfiguration.MenuPage> pageEntry, Integer pageNum) {
         this.skyShop = skyShop;
         this.menuManager = menuManager;
         this.shopManager = shopManager;
         this.inventoryManager = inventoryManager;
         this.localeManager = localeManager;
         this.pageEntry = pageEntry;
-        this.shopId = shopId;
         this.pageNum = pageNum;
 
         createInventory();
@@ -89,41 +77,43 @@ public class ShopGUI extends InventoryGUI {
      * A method to create the base structure of the inventory GUI.
     */
     public void createInventory() {
-        ShopConfiguration.ShopPage shopPage = pageEntry.getValue();
-        int shopSize = shopPage.size();
-        Component shopName = MiniMessage.miniMessage().deserialize(shopPage.name());
-        setInventory(Bukkit.createInventory(null, shopSize, shopName));
-        decorate();
+        MenuConfiguration.MenuPage page = pageEntry.getValue();
+        int menuSize = page.size();
+        Component menuName = MiniMessage.miniMessage().deserialize(page.name());
+        setInventory(Bukkit.createInventory(null, menuSize, menuName));
     }
 
     /**
      * A method to create all the buttons in the inventory GUI.
     */
     public void decorate() {
-        ShopConfiguration.ShopPage shopPage = pageEntry.getValue();
-        for (Map.Entry<String, ShopConfiguration.ShopEntry> stringShopEntryEntry : shopPage.entries().entrySet()) {
+        MenuConfiguration.MenuPage page = pageEntry.getValue();
+        for (Map.Entry<String, MenuConfiguration.MenuEntry> stringMenuEntryEntry : page.entries().entrySet()) {
             Material material;
-            List<Component> loreList;
+            List<Component> lore;
             Component name;
+            List<Component> list1;
             int i;
-            ShopConfiguration.ShopEntry entry = stringShopEntryEntry.getValue();
+            MenuConfiguration.MenuEntry itemEntry = stringMenuEntryEntry.getValue();
+            MenuConfiguration.Item item = itemEntry.item();
+            TransactionType type = TransactionType.valueOf(itemEntry.type());
 
-            ShopConfiguration.Item item = entry.item();
-            switch (TransactionType.valueOf(entry.type())) {
+            switch (type) {
                 case FILLER -> {
                     material = Material.valueOf(item.material());
                     name = MiniMessage.miniMessage().deserialize(item.name()).decorationIfAbsent(TextDecoration.ITALIC, TextDecoration.State.FALSE);
-                    loreList = new ArrayList<>();
+                    list1 = new ArrayList<>();
                     if (item.lore() != null) {
                         for (String loreLine : item.lore()) {
-                            loreList.add(MiniMessage.miniMessage().deserialize(loreLine).decorationIfAbsent(TextDecoration.ITALIC, TextDecoration.State.FALSE));
+                            list1.add(MiniMessage.miniMessage().deserialize(loreLine).decorationIfAbsent(TextDecoration.ITALIC, TextDecoration.State.FALSE));
                         }
                     }
 
-                    for (i = 0; i <= shopPage.size() - 1; i++) {
+                    for (i = 0; i <= page.size() - 1; i++) {
                         addButton(i, (new InventoryButton.Builder())
-                                .setItemStack(new ItemStack(material)).setItemName(name)
-                                .setLore(loreList)
+                                .setItemStack(new ItemStack(material))
+                                .setItemName(name)
+                                .setLore(list1)
                                 .setAction(event -> {
                                 })
                                 .build());
@@ -131,103 +121,91 @@ public class ShopGUI extends InventoryGUI {
                 }
 
                 case PREVIOUS_PAGE -> {
-                    loreList = new ArrayList<>();
+                    lore = new ArrayList<>();
                     if (item.lore() != null) {
                         for (String loreLine : item.lore()) {
-                            loreList.add(MiniMessage.miniMessage().deserialize(loreLine).decorationIfAbsent(TextDecoration.ITALIC, TextDecoration.State.FALSE));
+                            lore.add(MiniMessage.miniMessage().deserialize(loreLine).decorationIfAbsent(TextDecoration.ITALIC, TextDecoration.State.FALSE));
                         }
                     }
 
-                    addButton(entry.slot(), new InventoryButton.Builder()
+                    addButton(itemEntry.slot(), (new InventoryButton.Builder())
                             .setItemStack(new ItemStack(Material.valueOf(item.material())))
                             .setItemName(MiniMessage.miniMessage().deserialize(item.name()).decorationIfAbsent(TextDecoration.ITALIC, TextDecoration.State.FALSE))
-                            .setLore(loreList)
+                            .setLore(lore)
                             .setAction(event -> {
-                                List<Map.Entry<String, ShopConfiguration.ShopPage>> pageList = shopManager.getShopConfig(shopId).pages().entrySet().stream().toList();
+                                List<Map.Entry<String, MenuConfiguration.MenuPage>> pageList = menuManager.getMenuConfiguration().pages().entrySet().stream().toList();
                                 int prevPageNum = pageNum - 1;
-                                Map.Entry<String, ShopConfiguration.ShopPage> prevPage = pageList.get(prevPageNum);
+                                Map.Entry<String, MenuConfiguration.MenuPage> prevPage = pageList.get(prevPageNum);
+
                                 Bukkit.getScheduler().runTaskLater(skyShop, () -> {
                                     event.getWhoClicked().closeInventory(InventoryCloseEvent.Reason.OPEN_NEW);
-                                    inventoryManager.openGUI(new ShopGUI(skyShop, menuManager, shopManager, inventoryManager, localeManager, prevPage, shopId, prevPageNum), (Player) event.getWhoClicked());
+                                    inventoryManager.openGUI(new MenuGUI(skyShop, menuManager, shopManager, inventoryManager, localeManager, prevPage, prevPageNum), (Player) event.getWhoClicked());
                                 }, 1L);
                             }).build());
                 }
 
                 case NEXT_PAGE -> {
-                    loreList = new ArrayList<>();
+                    lore = new ArrayList<>();
                     if (item.lore() != null) {
                         for (String loreLine : item.lore()) {
-                            loreList.add(MiniMessage.miniMessage().deserialize(loreLine).decorationIfAbsent(TextDecoration.ITALIC, TextDecoration.State.FALSE));
+                            lore.add(MiniMessage.miniMessage().deserialize(loreLine).decorationIfAbsent(TextDecoration.ITALIC, TextDecoration.State.FALSE));
                         }
                     }
 
-                    addButton(entry.slot(), new InventoryButton.Builder()
+                    addButton(itemEntry.slot(), (new InventoryButton.Builder())
                             .setItemStack(new ItemStack(Material.valueOf(item.material())))
                             .setItemName(MiniMessage.miniMessage().deserialize(item.name()).decorationIfAbsent(TextDecoration.ITALIC, TextDecoration.State.FALSE))
-                            .setLore(loreList)
+                            .setLore(lore)
                             .setAction(event -> {
-                                List<Map.Entry<String, ShopConfiguration.ShopPage>> pageList = shopManager.getShopConfig(shopId).pages().entrySet().stream().toList();
+                                List<Map.Entry<String, MenuConfiguration.MenuPage>> pageList = menuManager.getMenuConfiguration().pages().entrySet().stream().toList();
                                 int nextPageNum = pageNum + 1;
-                                Map.Entry<String, ShopConfiguration.ShopPage> nextPage = pageList.get(nextPageNum);
+                                Map.Entry<String, MenuConfiguration.MenuPage> nextPage = pageList.get(nextPageNum);
+
                                 Bukkit.getScheduler().runTaskLater(skyShop, () -> {
                                     event.getWhoClicked().closeInventory(InventoryCloseEvent.Reason.OPEN_NEW);
-                                    inventoryManager.openGUI(new ShopGUI(skyShop, menuManager, shopManager, inventoryManager, localeManager, nextPage, shopId, nextPageNum), (Player) event.getWhoClicked());
+                                    inventoryManager.openGUI(new MenuGUI(skyShop, menuManager, shopManager, inventoryManager, localeManager, nextPage, nextPageNum), (Player) event.getWhoClicked());
                                 }, 1L);
                             }).build());
                 }
 
                 case RETURN -> {
-                    loreList = new ArrayList<>();
+                    lore = new ArrayList<>();
                     if (item.lore() != null) {
                         for (String loreLine : item.lore()) {
-                            loreList.add(MiniMessage.miniMessage().deserialize(loreLine).decorationIfAbsent(TextDecoration.ITALIC, TextDecoration.State.FALSE));
+                            lore.add(MiniMessage.miniMessage().deserialize(loreLine).decorationIfAbsent(TextDecoration.ITALIC, TextDecoration.State.FALSE));
                         }
                     }
 
-                    addButton(entry.slot(), (new InventoryButton.Builder())
+                    addButton(itemEntry.slot(), (new InventoryButton.Builder())
                             .setItemStack(new ItemStack(Material.valueOf(item.material())))
                             .setItemName(MiniMessage.miniMessage().deserialize(item.name()).decorationIfAbsent(TextDecoration.ITALIC, TextDecoration.State.FALSE))
-                            .setLore(loreList)
-                            .setAction(event -> Bukkit.getScheduler().runTaskLater(skyShop, () -> {
-                                event.getWhoClicked().closeInventory(InventoryCloseEvent.Reason.OPEN_NEW);
-                                inventoryManager.openGUI(new MenuGUI(skyShop, menuManager, shopManager, inventoryManager, localeManager, menuManager.getMenuConfiguration().pages().entrySet().stream().toList().getFirst(), 0), (Player) event.getWhoClicked());
-                            }, 1L))
+                            .setLore(lore)
+                            .setAction(event -> Bukkit.getScheduler().runTaskLater(skyShop, () -> event.getWhoClicked().closeInventory(InventoryCloseEvent.Reason.OPEN_NEW), 1L))
                             .build());
                 }
 
-                case ITEM, COMMAND -> {
-                    loreList = new ArrayList<>();
+                case OPEN_SHOP -> {
+                    lore = new ArrayList<>();
                     if (item.lore() != null) {
                         for (String loreLine : item.lore()) {
-                            loreList.add(MiniMessage.miniMessage().deserialize(loreLine, new TagResolver[]{
-                                    Placeholder.parsed("buy_price", String.valueOf(entry.prices().buyPrice())),
-                                    Placeholder.parsed("sell_price", String.valueOf(entry.prices().sellPrice()))
-                            }).decorationIfAbsent(TextDecoration.ITALIC, TextDecoration.State.FALSE));
+                            lore.add(MiniMessage.miniMessage().deserialize(loreLine).decorationIfAbsent(TextDecoration.ITALIC, TextDecoration.State.FALSE));
                         }
                     }
 
-                    addButton(entry.slot(), new InventoryButton.Builder()
+                    addButton(itemEntry.slot(), (new InventoryButton.Builder())
                             .setItemStack(new ItemStack(Material.valueOf(item.material())))
                             .setItemName(MiniMessage.miniMessage().deserialize(item.name()).decorationIfAbsent(TextDecoration.ITALIC, TextDecoration.State.FALSE))
-                            .setLore(loreList)
+                            .setLore(lore)
                             .setAction(event -> Bukkit.getScheduler().runTaskLater(skyShop, () -> {
                                 event.getWhoClicked().closeInventory(InventoryCloseEvent.Reason.OPEN_NEW);
-                                inventoryManager.openGUI(new TransactionGUI(skyShop, menuManager, inventoryManager, shopManager, localeManager, pageEntry, entry, shopId, pageNum, 1), (Player) event.getWhoClicked());
+                                inventoryManager.openGUI(new ShopGUI(skyShop, menuManager, shopManager, inventoryManager, localeManager, shopManager.getShopConfig(itemEntry.shop()).pages().entrySet().stream().toList().get(0), itemEntry.shop(), 0), (Player) event.getWhoClicked());
                             }, 1L))
+
+
                             .build());
                 }
             }
         }
-
         super.decorate();
-    }
-
-    /**
-     * Handles the closing of the inventory GUI.
-     * @param event The InventoryCloseEvent
-    */
-    public void onClose(InventoryCloseEvent event) {
-        if (!event.getReason().equals(InventoryCloseEvent.Reason.OPEN_NEW) && !event.getReason().equals(InventoryCloseEvent.Reason.UNLOADED))
-            Bukkit.getScheduler().runTaskLater(skyShop, () -> inventoryManager.openGUI(new MenuGUI(skyShop, menuManager, shopManager, inventoryManager, localeManager, menuManager.getMenuConfiguration().pages().entrySet().stream().toList().getFirst(), 0), (Player)event.getPlayer()), 1L);
     }
 }
