+++ Core/src/main/java/com/github/lukesky19/skyshop/util/PlaceholderAPIUtil.java
