+++ Core/src/main/java/com/github/lukesky19/skyshop/util/enums/TransactionType.java
