+++ Core/src/main/resources/menu.yml
