+++ Core/src/main/resources/plugin.yml
@@ -1,7 +1,7 @@
 name: SkyShop
 version: '${version}'
 main: com.github.lukesky19.skyshop.SkyShop
-api-version: '1.20'
+api-version: '1.17'
 load: STARTUP
 author: lukeskywlker19
 description: A graphical shop plugin
