+++ Core/src/main/resources/shops/example.yml
