+++ settings.gradle
@@ -1,1 +1,2 @@
 rootProject.name = 'SkyShop'
+include 'Core', 'v1_20_R4'
