+++ v1_20_R4/build.gradle
@@ -1,0 +1,4 @@
+dependencies {
+    compileOnly 'io.papermc.paper:paper-api:1.20.6-R0.1-SNAPSHOT'
+    implementation project(':Core')
+}
