+++ v1_20_R4/gradle.properties
@@ -1,0 +1,2 @@
+# Valid for Minecraft version 1.20.6(v1_20_R4) and 1.21 (v1_21_R1)
+javaVersion=21
